requirements.txt
pygame>=2.0.0
pytmx>=3.21.7
pyserial>=3.5
psutil>=5.9.0  # Optional: Memory monitoring für RPi Tests (F9)
Pillow>=9.0.0  # GIF-Verarbeitung für Easter Eggs
opencv-python-headless>=4.5.0  # Video-Wiedergabe für Cinematics (headless = kein SDL-Konflikt mit pygame)
